refactor: self-invocation 문제 해결을 위해 강의 동기화 구조 분리
